Fix: BigDecimal en vez de double

+++ src/main/java/dds/monedero/model/Movimiento.java
@@ -1,15 +1,14 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Movimiento {
   private LocalDate fecha;
-  //En ningún lenguaje de programación usen jamás doubles para modelar dinero en el mundo real
-  //siempre usen numeros de precision arbitraria, como BigDecimal en Java y similares
-  private double monto;
+  private BigDecimal monto;
   private TipoMovimiento tipoMovimiento;
 
-  public Movimiento(LocalDate fecha, double monto, TipoMovimiento tipoMovimiento) {
+  public Movimiento(LocalDate fecha, BigDecimal monto, TipoMovimiento tipoMovimiento) {
     this.fecha = fecha;
     this.monto = monto;
     this.tipoMovimiento = tipoMovimiento;
@@ -21,8 +20,8 @@ public class Movimiento {
 
   public boolean esDeLaFecha(LocalDate fecha) { return this.fecha.equals(fecha); }
 
-  public double calcularValor() { return tipoMovimiento.calcularValor(getMonto()); }
+  public BigDecimal calcularValor() { return tipoMovimiento.calcularValor(getMonto()); }
 
-  public double getMonto() {return monto;}
+  public BigDecimal getMonto() {return monto;}
   public TipoMovimiento getTipoMovimiento(){return tipoMovimiento;}
 }
